+++ ear/fileio/adm/elements/__init__.py
@@ -1,12 +1,12 @@
 # flake8: noqa
 from .main_elements import (AudioChannelFormat, AudioPackFormat, AudioTrackFormat,
                             AudioStreamFormat, AudioProgramme, AudioContent,
-                            AudioObject, AudioObjectInteraction, AudioTrackUID,
-                            FormatDefinition, GainInteractionRange, PositionInteractionRange,
-                            TypeDefinition, Frequency)
+                            AudioObject, AudioTrackUID, FormatDefinition, TypeDefinition,
+                            Frequency)
 from .block_formats import (AudioBlockFormatObjects, ChannelLock, ObjectDivergence,
                             JumpPosition, AudioBlockFormatDirectSpeakers, AudioBlockFormatBinaural, 
                             AudioBlockFormatHoa, AudioBlockFormatMatrix, MatrixCoefficient,
                             CartesianZone, PolarZone)
 from .geom import (DirectSpeakerPolarPosition, DirectSpeakerCartesianPosition, BoundCoordinate,
                    ObjectPolarPosition, ObjectCartesianPosition, ScreenEdgeLock)
+from .interaction import (AudioObjectInteraction, GainInteractionRange, PositionInteractionRange)
