+++ ear/fileio/adm/elements/main_elements.py
@@ -6,6 +6,7 @@ from six import string_types
 
 from ..exceptions import AdmError
 from ....common import CartesianScreen, PolarScreen, default_screen, list_of
+from .interaction import AudioObjectInteraction
 
 
 def _lookup_elements(adm, idRefs):
@@ -82,34 +83,6 @@ class AudioContent(ADMElement):
         if self.audioObjectIDRef is not None:
             self.audioObjects = _lookup_elements(adm, self.audioObjectIDRef)
             self.audioObjectIDRef = None
-
-@attrs(slots=True)
-class PositionInteractionRange(object):
-    azimuthMin = attrib(default=None, validator=optional(instance_of(float)))
-    azimuthMax = attrib(default=None, validator=optional(instance_of(float)))
-    elevationMin = attrib(default=None, validator=optional(instance_of(float)))
-    elevationMax = attrib(default=None, validator=optional(instance_of(float)))
-    distanceMin = attrib(default=None, validator=optional(instance_of(float)))
-    distanceMax = attrib(default=None, validator=optional(instance_of(float)))
-    XMin = attrib(default=None, validator=optional(instance_of(float)))
-    XMax = attrib(default=None, validator=optional(instance_of(float)))
-    YMin = attrib(default=None, validator=optional(instance_of(float)))
-    YMax = attrib(default=None, validator=optional(instance_of(float)))
-    ZMin = attrib(default=None, validator=optional(instance_of(float)))
-    ZMax = attrib(default=None, validator=optional(instance_of(float)))
-
-@attrs(slots=True)
-class GainInteractionRange(object):
-    min = attrib(default=None, validator=optional(instance_of(float)))
-    max = attrib(default=None, validator=optional(instance_of(float)))
-
-@attrs(slots=True)
-class AudioObjectInteraction(object):
-    onOffInteract = attrib(default=None, validator=optional(instance_of(int)))
-    gainInteract = attrib(default=None, validator=optional(instance_of(int)))
-    positionInteract = attrib(default=None, validator=optional(instance_of(int)))
-    gainInteractionRange = attrib(default=None)
-    positionInteractionRange = attrib(default=None)
 
 @attrs(slots=True)
 class AudioObject(ADMElement):
